+++ src/modules/Playerbot/Combat/CombatEventBus.cpp
@@ -12,181 +12,9 @@
 #include "Player.h"
 #include "Log.h"
 #include "Timer.h"
-#include <sstream>
 
 namespace Playerbot
 {
-
-// ============================================================================
-// CombatEvent Helper Constructors
-// ============================================================================
-
-CombatEvent CombatEvent::SpellCastStart(ObjectGuid caster, ObjectGuid target, uint32 spellId, uint32 castTime)
-{
-    CombatEvent event;
-    event.type = CombatEventType::SPELL_CAST_START;
-    event.priority = CombatEventPriority::CRITICAL;
-    event.casterGuid = caster;
-    event.targetGuid = target;
-    event.victimGuid = target;
-    event.spellId = spellId;
-    event.amount = static_cast<int32>(castTime);
-    event.schoolMask = 0;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-CombatEvent CombatEvent::SpellCastGo(ObjectGuid caster, ObjectGuid target, uint32 spellId)
-{
-    CombatEvent event;
-    event.type = CombatEventType::SPELL_CAST_GO;
-    event.priority = CombatEventPriority::HIGH;
-    event.casterGuid = caster;
-    event.targetGuid = target;
-    event.victimGuid = target;
-    event.spellId = spellId;
-    event.amount = 0;
-    event.schoolMask = 0;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-CombatEvent CombatEvent::SpellDamage(ObjectGuid caster, ObjectGuid victim, uint32 spellId, int32 damage, uint32 school)
-{
-    CombatEvent event;
-    event.type = CombatEventType::SPELL_DAMAGE_DEALT;
-    event.priority = CombatEventPriority::HIGH;
-    event.casterGuid = caster;
-    event.targetGuid = victim;
-    event.victimGuid = victim;
-    event.spellId = spellId;
-    event.amount = damage;
-    event.schoolMask = school;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-CombatEvent CombatEvent::SpellHeal(ObjectGuid caster, ObjectGuid target, uint32 spellId, int32 heal)
-{
-    CombatEvent event;
-    event.type = CombatEventType::SPELL_HEAL_DEALT;
-    event.priority = CombatEventPriority::HIGH;
-    event.casterGuid = caster;
-    event.targetGuid = target;
-    event.victimGuid = target;
-    event.spellId = spellId;
-    event.amount = heal;
-    event.schoolMask = 0;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-CombatEvent CombatEvent::SpellInterrupt(ObjectGuid interrupter, ObjectGuid victim, uint32 interruptedSpell, uint32 interruptSpell)
-{
-    CombatEvent event;
-    event.type = CombatEventType::SPELL_INTERRUPTED;
-    event.priority = CombatEventPriority::CRITICAL;
-    event.casterGuid = interrupter;
-    event.targetGuid = victim;
-    event.victimGuid = victim;
-    event.spellId = interruptedSpell;
-    event.amount = static_cast<int32>(interruptSpell);
-    event.schoolMask = 0;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-CombatEvent CombatEvent::AttackStart(ObjectGuid attacker, ObjectGuid victim)
-{
-    CombatEvent event;
-    event.type = CombatEventType::ATTACK_START;
-    event.priority = CombatEventPriority::HIGH;
-    event.casterGuid = attacker;
-    event.targetGuid = victim;
-    event.victimGuid = victim;
-    event.spellId = 0;
-    event.amount = 0;
-    event.schoolMask = 0;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(10000);
-    return event;
-}
-
-CombatEvent CombatEvent::AttackStop(ObjectGuid attacker, ObjectGuid victim, bool nowDead)
-{
-    CombatEvent event;
-    event.type = CombatEventType::ATTACK_STOP;
-    event.priority = CombatEventPriority::HIGH;
-    event.casterGuid = attacker;
-    event.targetGuid = victim;
-    event.victimGuid = victim;
-    event.spellId = 0;
-    event.amount = nowDead ? 1 : 0;
-    event.schoolMask = 0;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-CombatEvent CombatEvent::ThreatUpdate(ObjectGuid unit, ObjectGuid victim, int32 threatChange)
-{
-    CombatEvent event;
-    event.type = CombatEventType::THREAT_UPDATE;
-    event.priority = CombatEventPriority::MEDIUM;
-    event.casterGuid = unit;
-    event.targetGuid = victim;
-    event.victimGuid = victim;
-    event.spellId = 0;
-    event.amount = threatChange;
-    event.schoolMask = 0;
-    event.flags = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-bool CombatEvent::IsValid() const
-{
-    if (type >= CombatEventType::MAX_COMBAT_EVENT)
-        return false;
-    if (timestamp.time_since_epoch().count() == 0)
-        return false;
-    if (casterGuid.IsEmpty() && targetGuid.IsEmpty())
-        return false;
-    return true;
-}
-
-bool CombatEvent::IsExpired() const
-{
-    return std::chrono::steady_clock::now() >= expiryTime;
-}
-
-std::string CombatEvent::ToString() const
-{
-    std::ostringstream oss;
-    oss << "[CombatEvent] Type: " << static_cast<uint32>(type)
-        << ", Caster: " << casterGuid.ToString()
-        << ", Target: " << targetGuid.ToString()
-        << ", Spell: " << spellId
-        << ", Amount: " << amount;
-    return oss.str();
-}
-
-// ============================================================================
-// CombatEventBus Implementation (Minimal)
-// ============================================================================
 
 CombatEventBus::CombatEventBus()
 {
