+++ src/modules/Playerbot/Cooldown/CooldownEventBus.cpp
@@ -12,92 +12,9 @@
 #include "Player.h"
 #include "Log.h"
 #include "Timer.h"
-#include <sstream>
 
 namespace Playerbot
 {
-
-// ============================================================================
-// CooldownEvent Helper Constructors
-// ============================================================================
-
-CooldownEvent CooldownEvent::SpellCooldownStart(ObjectGuid caster, uint32 spellId, uint32 cooldownMs)
-{
-    CooldownEvent event;
-    event.type = CooldownEventType::SPELL_COOLDOWN_START;
-    event.priority = CooldownEventPriority::MEDIUM;
-    event.casterGuid = caster;
-    event.spellId = spellId;
-    event.itemId = 0;
-    event.category = 0;
-    event.cooldownMs = cooldownMs;
-    event.modRateMs = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(cooldownMs + 5000);
-    return event;
-}
-
-CooldownEvent CooldownEvent::SpellCooldownClear(ObjectGuid caster, uint32 spellId)
-{
-    CooldownEvent event;
-    event.type = CooldownEventType::SPELL_COOLDOWN_CLEAR;
-    event.priority = CooldownEventPriority::HIGH;
-    event.casterGuid = caster;
-    event.spellId = spellId;
-    event.itemId = 0;
-    event.category = 0;
-    event.cooldownMs = 0;
-    event.modRateMs = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-CooldownEvent CooldownEvent::ItemCooldownStart(ObjectGuid caster, uint32 itemId, uint32 cooldownMs)
-{
-    CooldownEvent event;
-    event.type = CooldownEventType::ITEM_COOLDOWN_START;
-    event.priority = CooldownEventPriority::MEDIUM;
-    event.casterGuid = caster;
-    event.spellId = 0;
-    event.itemId = itemId;
-    event.category = 0;
-    event.cooldownMs = cooldownMs;
-    event.modRateMs = 0;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(cooldownMs + 5000);
-    return event;
-}
-
-bool CooldownEvent::IsValid() const
-{
-    if (type >= CooldownEventType::MAX_COOLDOWN_EVENT)
-        return false;
-    if (timestamp.time_since_epoch().count() == 0)
-        return false;
-    if (casterGuid.IsEmpty())
-        return false;
-    return true;
-}
-
-bool CooldownEvent::IsExpired() const
-{
-    return std::chrono::steady_clock::now() >= expiryTime;
-}
-
-std::string CooldownEvent::ToString() const
-{
-    std::ostringstream oss;
-    oss << "[CooldownEvent] Type: " << static_cast<uint32>(type)
-        << ", Caster: " << casterGuid.ToString()
-        << ", Spell: " << spellId
-        << ", Item: " << itemId
-        << ", Duration: " << cooldownMs << "ms";
-    return oss.str();
-}
-
-// ============================================================================
-// CooldownEventBus Implementation
 // ============================================================================
 
 CooldownEventBus::CooldownEventBus()
