+++ src/modules/Playerbot/Loot/LootEventBus.h
@@ -12,10 +12,8 @@
 
 #include "Define.h"
 #include "Threading/LockHierarchy.h"
-#include "ObjectGuid.h"
+#include "LootEvents.h"
 #include "Core/DI/Interfaces/ILootEventBus.h"
-#include <chrono>
-#include <string>
 #include <vector>
 #include <unordered_map>
 #include <queue>
@@ -26,73 +24,6 @@ namespace Playerbot
 {
 
 class BotAI;
-
-enum class LootEventType : uint8
-{
-    LOOT_WINDOW_OPENED = 0,
-    LOOT_WINDOW_CLOSED,
-    LOOT_ITEM_RECEIVED,
-    LOOT_MONEY_RECEIVED,
-    LOOT_REMOVED,
-    LOOT_SLOT_CHANGED,
-    LOOT_ROLL_STARTED,
-    LOOT_ROLL_CAST,
-    LOOT_ROLL_WON,
-    LOOT_ALL_PASSED,
-    MASTER_LOOT_LIST,
-    MAX_LOOT_EVENT
-};
-
-enum class LootEventPriority : uint8
-{
-    CRITICAL = 0,
-    HIGH = 1,
-    MEDIUM = 2,
-    LOW = 3,
-    BATCH = 4
-};
-
-enum class LootType : uint8
-{
-    CORPSE = 0,
-    PICKPOCKETING,
-    FISHING,
-    DISENCHANTING,
-    SKINNING,
-    PROSPECTING,
-    MILLING,
-    ITEM,
-    MAIL,
-    INSIGNIA
-};
-
-struct LootEvent
-{
-    LootEventType type;
-    LootEventPriority priority;
-    ObjectGuid looterGuid;
-    ObjectGuid itemGuid;
-    uint32 itemEntry;
-    uint32 itemCount;
-    LootType lootType;
-    std::chrono::steady_clock::time_point timestamp;
-    std::chrono::steady_clock::time_point expiryTime;
-
-    bool IsValid() const;
-    bool IsExpired() const;
-    std::string ToString() const;
-
-    // Helper constructors
-    static LootEvent ItemLooted(ObjectGuid looter, ObjectGuid item, uint32 entry, uint32 count, LootType type);
-    static LootEvent LootRollStarted(ObjectGuid item, uint32 entry);
-    static LootEvent LootRollWon(ObjectGuid winner, ObjectGuid item, uint32 entry);
-
-    // Priority comparison for priority queue
-    bool operator<(LootEvent const& other) const
-    {
-        return priority > other.priority;
-    }
-};
 
 class TC_GAME_API LootEventBus final : public ILootEventBus
 {
